+++ src/custom/pages/Account/Tokens/TokensOverview.tsx
@@ -7,13 +7,10 @@ import {
   MenuItem,
   MenuWrapper,
   StyledChevronDown,
-  Wrapper,
-  AccountPageWrapper,
   Subtitle,
   AccountHeading,
   RemoveTokens,
 } from './styled'
-import { AccountMenu } from '../Menu'
 import { useAllTokens } from 'hooks/Tokens'
 import { isTruthy } from 'utils/misc'
 import TokensTable from 'components/Tokens/TokensTable'
@@ -24,6 +21,7 @@ import { useOnClickOutside } from 'hooks/useOnClickOutside'
 import useTheme from 'hooks/useTheme'
 import usePrevious from 'hooks/usePrevious'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import WrapperWithNav from '../WrapperWithNav'
 
 export enum PageViewKeys {
   ALL_TOKENS = 'ALL_TOKENS',
@@ -105,43 +103,36 @@ export default function TokensOverview() {
   }, [chainId, prevChainId, prevSelectedView, selectedView])
 
   return (
-    <Wrapper>
-      <AccountMenu />
-      <AccountPageWrapper>
-        <AccountHeading>
-          <MenuWrapper ref={node as any}>
-            <MenuButton onClick={toggleMenu}>
-              <Subtitle>
-                <Trans>{PageView[selectedView].label}</Trans>
-              </Subtitle>
-              <StyledChevronDown size={14} />
-            </MenuButton>
+    <WrapperWithNav name="Tokens" id="account-tokens">
+      <AccountHeading>
+        <MenuWrapper ref={node as any}>
+          <MenuButton onClick={toggleMenu}>
+            <Subtitle>
+              <Trans>{PageView[selectedView].label}</Trans>
+            </Subtitle>
+            <StyledChevronDown size={14} />
+          </MenuButton>
 
-            {isMenuOpen ? (
-              <Menu>
-                {Object.entries(PageView).map(([key, value]) => (
-                  <MenuItem
-                    key={key}
-                    active={selectedView === key}
-                    onClick={() => handleMenuClick(key as PageViewKeys)}
-                  >
-                    <span>{value.label}</span>
-                    {selectedView === key ? <Check size={20} color={theme.green1} /> : null}
-                  </MenuItem>
-                ))}
-              </Menu>
-            ) : null}
-          </MenuWrapper>
+          {isMenuOpen ? (
+            <Menu>
+              {Object.entries(PageView).map(([key, value]) => (
+                <MenuItem key={key} active={selectedView === key} onClick={() => handleMenuClick(key as PageViewKeys)}>
+                  <span>{value.label}</span>
+                  {selectedView === key ? <Check size={20} color={theme.green1} /> : null}
+                </MenuItem>
+              ))}
+            </Menu>
+          ) : null}
+        </MenuWrapper>
 
-          {selectedView === PageViewKeys.FAVORITE_TOKENS && (
-            <RemoveTokens onClick={handleRestoreTokens}>
-              (<Trans>Restore defaults</Trans>)
-            </RemoveTokens>
-          )}
-        </AccountHeading>
+        {selectedView === PageViewKeys.FAVORITE_TOKENS && (
+          <RemoveTokens onClick={handleRestoreTokens}>
+            (<Trans>Restore defaults</Trans>)
+          </RemoveTokens>
+        )}
+      </AccountHeading>
 
-        {renderTableContent()}
-      </AccountPageWrapper>
-    </Wrapper>
+      {renderTableContent()}
+    </WrapperWithNav>
   )
 }
